fix(ci): python-full-coverage — test stubs missing agent_runtime_executor seam + part_013 over line limit
Commit 864cf2a7e added an injectable agent_runtime_executor seam to
WorkspaceExecutor and the get_adapter call site, but the test stubs
did not follow:

- _Executor stubs in test_executor_mirror_hooks_path{,_commit}.py
  lacked the _agent_runtime_executor attribute, so get_adapter raised
  AttributeError before the mirror-repair warning could be logged
  (surfacing as executor.unexpected_in_agent instead). The mocked
  adapter SimpleNamespace also lacked is_hosted, which the hosted-mode
  _should_skip_compose_recovery guard now reads.
- _WorkspaceExecutor stubs in test_worker.py and
  test_worker_runtime_wiring.py lacked the agent_runtime_executor
  kwarg, so worker.py's WorkspaceExecutor(...) call raised TypeError.
- test_pr_monitor_runner_part_013.py grew to 1516 lines (limit 1500);
  split the cleanup-service-down recovery tests into part_016.
