+++ tests/unit/runtime/test_pr_monitor_runner_parts/test_pr_monitor_runner_part_013.py
@@ -25,7 +25,6 @@ from awf.runtime.pr_monitor_runner.types import (
     _MonitorAgentRuntimeOwnershipRepairFailedError,
     _MonitorAgentServiceRecoveryFailedError,
     _MonitorAgentServiceRecoverySupersededError,
-    _MonitorHeadObjectMissingError,
     _MonitorMirrorHooksPathRepairFailedError,
 )
 from tests.postgres import postgres_test_engine
@@ -741,776 +740,3 @@ async def test_monitor_agent_restart_timeout_invalid_profile_uses_default(
     )
 
     assert timeout == 300
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_restarts_service_and_retries(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: cleanup restarted")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    probe = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        return_value=True,
-    )
-    command_evidence: list[str] = []
-    compose_file = tmp_path / "compose.yml"
-
-    result = await runner._run_monitor_agent_with_service_recovery(
-        workspace_id=workspace_id,
-        compose_project="proj",
-        compose_file=compose_file,
-        prompt="fix the comment",
-        log_source="recovery",
-        command_evidence=command_evidence,
-    )
-
-    assert result.stdout == "AWF-VERDICT: FIXED: cleanup restarted"
-    assert adapter.calls == ["fix the comment", "fix the comment"]
-    assert command_evidence == [
-        'service "agent" is not running',
-        "AWF-VERDICT: FIXED: cleanup restarted",
-    ]
-    probe.assert_awaited_once()
-    ensure_project_up.assert_awaited_once_with(
-        project_name="proj",
-        compose_file=compose_file,
-        workspace_id=workspace_id,
-        wait=True,
-        compose_up_timeout_seconds=300,
-        force_recreate=True,
-        services=(),
-    )
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_repairs_git_before_restart(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    calls: list[str] = []
-
-    class RecordingAdapter(FakeAdapter):
-        async def run(self, **kwargs: object):  # type: ignore[no-untyped-def,override]
-            calls.append("adapter.run")
-            return await super().run(**kwargs)  # type: ignore[arg-type]
-
-    adapter = RecordingAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: cleanup repaired")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    mirror_path = tmp_path / "mirror.git"
-
-    async def _ensure_project_up(**_kwargs: object) -> None:
-        calls.append("ensure_project_up")
-
-    async def _repair_mirror_hooks_path(_mirror_path: Path) -> bool:
-        calls.append("mirror_hooks_repair")
-        return True
-
-    async def _verify_head_object_exists(_worktree_path: Path) -> bool:
-        calls.append("verify_head")
-        return False
-
-    async def _open_merge_candidate_head_sha(_workspace_id: str) -> str:
-        calls.append("open_candidate_head")
-        return "abc123"
-
-    async def _recover_missing_head_object_from_filesystem(
-        *_args: object,
-        **_kwargs: object,
-    ) -> str:
-        calls.append("recover_missing_head")
-        return "def456"
-
-    async def _repair_agent_runtime_ownership(**_kwargs: object) -> bool:
-        calls.append("ownership_repair")
-        return True
-
-    runner._open_merge_candidate_head_sha = _open_merge_candidate_head_sha  # type: ignore[method-assign]
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        side_effect=_ensure_project_up,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.mirror_path_for_worktree",
-        return_value=mirror_path,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.repair_mirror_hooks_path",
-        side_effect=_repair_mirror_hooks_path,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        side_effect=_verify_head_object_exists,
-        create=True,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery._recover_missing_head_object_from_filesystem",
-        side_effect=_recover_missing_head_object_from_filesystem,
-        create=True,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.repair_agent_runtime_ownership",
-        side_effect=_repair_agent_runtime_ownership,
-    )
-
-    result = await runner._run_monitor_agent_with_service_recovery(
-        workspace_id=workspace_id,
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        prompt="fix the comment",
-        log_source="recovery",
-        command_evidence=[],
-    )
-
-    assert result.stdout == "AWF-VERDICT: FIXED: cleanup repaired"
-    assert calls == [
-        "adapter.run",
-        "mirror_hooks_repair",
-        "verify_head",
-        "open_candidate_head",
-        "recover_missing_head",
-        "ensure_project_up",
-        "ownership_repair",
-        "mirror_hooks_repair",
-        "adapter.run",
-    ]
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_checks_claim_before_git_repair(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    async with factory() as session:
-        workspace = await WorkspaceRepository(session).get(workspace_id)
-        assert workspace is not None
-        workspace.monitor_claimed_by = "worker-new"
-        await session.commit()
-
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: should not run")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    runner._monitor_owner_id = "worker-old"
-
-    async def _repair_mirror_hooks_path(_mirror_path: Path) -> bool:
-        pytest.fail("superseded cleanup recovery must not repair git")
-
-    async def _verify_head_object_exists(_worktree_path: Path) -> bool:
-        pytest.fail("superseded cleanup recovery must not inspect HEAD")
-
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    repair_mirror_hooks_path = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.repair_mirror_hooks_path",
-        side_effect=_repair_mirror_hooks_path,
-    )
-    verify_head_object_exists = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        side_effect=_verify_head_object_exists,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.mirror_path_for_worktree",
-        return_value=tmp_path / "mirror.git",
-    )
-    compose_file = _write_compose_file(tmp_path)
-
-    with pytest.raises(_MonitorAgentServiceRecoverySupersededError) as raised:
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=compose_file,
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    repair_mirror_hooks_path.assert_not_awaited()
-    verify_head_object_exists.assert_not_awaited()
-    ensure_project_up.assert_not_awaited()
-    assert raised.value.details["superseded_reason"] == "monitor_claim_changed"
-    assert raised.value.details["restart_attempts"] == 1
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_stops_when_missing_head_repair_fails(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-
-    async def _open_merge_candidate_head_sha(_workspace_id: str) -> str:
-        return "abc123"
-
-    runner._open_merge_candidate_head_sha = _open_merge_candidate_head_sha  # type: ignore[method-assign]
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.mirror_path_for_worktree",
-        return_value=tmp_path / "mirror.git",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.repair_mirror_hooks_path",
-        return_value=True,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        return_value=False,
-        create=True,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery._recover_missing_head_object_from_filesystem",
-        return_value=None,
-        create=True,
-    )
-
-    with pytest.raises(_MonitorHeadObjectMissingError):
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=tmp_path / "compose.yml",
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_not_awaited()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_stops_when_mirror_repair_fails(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    repair_error = GitOperationError(
-        operation="mirror.hooks_path_repair",
-        returncode=1,
-        stdout="",
-        stderr="fatal: could not unset hooksPath",
-        reason_code="MIRROR_HOOKS_PATH_REPAIR_FAILED",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.mirror_path_for_worktree",
-        return_value=tmp_path / "mirror.git",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.repair_mirror_hooks_path",
-        side_effect=repair_error,
-    )
-
-    with pytest.raises(_MonitorMirrorHooksPathRepairFailedError):
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=tmp_path / "compose.yml",
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_not_awaited()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_stops_when_no_recovery_head_exists(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr='service "agent" is not running',
-            ),
-        )
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-
-    async def _open_merge_candidate_head_sha(_workspace_id: str) -> None:
-        return None
-
-    runner._open_merge_candidate_head_sha = _open_merge_candidate_head_sha  # type: ignore[method-assign]
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.mirror_path_for_worktree",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        return_value=False,
-        create=True,
-    )
-
-    with pytest.raises(_MonitorHeadObjectMissingError) as raised:
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=tmp_path / "compose.yml",
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert raised.value.reason_code == "HEAD_OBJECT_MISSING_UNRECOVERABLE"
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_not_awaited()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_cleanup_service_down_uses_exception_message_when_output_empty(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=ComposeExecCleanupError(
-            invocation_id="awf-test-cleanup",
-            source="agent",
-            label="monitor",
-            message='service "agent" is not running',
-            cleanup_result=CommandResult(returncode=1, stdout="", stderr=""),
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: cleanup restarted")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.verify_head_object_exists",
-        return_value=True,
-    )
-
-    result = await runner._run_monitor_agent_with_service_recovery(
-        workspace_id=workspace_id,
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        prompt="fix the comment",
-        log_source="recovery",
-        command_evidence=[],
-    )
-
-    assert result.stdout == "AWF-VERDICT: FIXED: cleanup restarted"
-    assert adapter.calls == ["fix the comment", "fix the comment"]
-    ensure_project_up.assert_awaited_once()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_service_recovery_stops_when_workspace_leaves_monitoring(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=AgentRunError(
-            agent=AgentRuntime.claude_code,
-            result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr="monitor idle timeout while agent service was down",
-            ),
-            reason_code=AGENT_IDLE_TIMEOUT,
-            details={"provider": "google", "model": "gemini-2.5-pro"},
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: should not run")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-
-    async def _cancel_workspace_after_restart(*_args: object, **_kwargs: object) -> None:
-        async with factory() as session:
-            repo = WorkspaceRepository(session)
-            workspace = await repo.get(workspace_id)
-            assert workspace is not None
-            await repo.transition(
-                workspace,
-                to=WorkspaceStatus.cancelled,
-                reason_code="TEST_CANCELLED_DURING_RESTART",
-            )
-            await session.commit()
-
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        side_effect=_cancel_workspace_after_restart,
-    )
-    compose_file = _write_compose_file(tmp_path)
-
-    with pytest.raises(_MonitorAgentServiceRecoverySupersededError):
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=compose_file,
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_awaited_once()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_service_recovery_checks_monitor_claim_before_restart(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    async with factory() as session:
-        workspace = await WorkspaceRepository(session).get(workspace_id)
-        assert workspace is not None
-        workspace.monitor_claimed_by = "worker-new"
-        await session.commit()
-
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=AgentRunError(
-            agent=AgentRuntime.claude_code,
-            result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr="monitor idle timeout while agent service was down",
-            ),
-            reason_code=AGENT_IDLE_TIMEOUT,
-            details={"provider": "google", "model": "gemini-2.5-pro"},
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: should not run")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    runner._monitor_owner_id = "worker-old"
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-    compose_file = _write_compose_file(tmp_path)
-
-    with pytest.raises(_MonitorAgentServiceRecoverySupersededError) as raised:
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=compose_file,
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_not_awaited()
-    assert raised.value.details["superseded_reason"] == "monitor_claim_changed"
-    assert raised.value.details["restart_attempts"] == 1
-
-
-@pytest.mark.unit
-async def test_monitor_agent_service_recovery_stops_when_monitor_claim_is_superseded(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    async with factory() as session:
-        workspace = await WorkspaceRepository(session).get(workspace_id)
-        assert workspace is not None
-        workspace.monitor_claimed_by = "worker-old"
-        await session.commit()
-
-    adapter = FakeAdapter()
-    adapter.queue(
-        exc=AgentRunError(
-            agent=AgentRuntime.claude_code,
-            result=CommandResult(
-                returncode=1,
-                stdout="",
-                stderr="monitor idle timeout while agent service was down",
-            ),
-            reason_code=AGENT_IDLE_TIMEOUT,
-            details={"provider": "google", "model": "gemini-2.5-pro"},
-        )
-    )
-    adapter.queue(stdout="AWF-VERDICT: FIXED: should not run")
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    runner._monitor_owner_id = "worker-old"
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-
-    async def _supersede_monitor_claim_after_restart(*_args: object, **_kwargs: object) -> None:
-        async with factory() as session:
-            workspace = await WorkspaceRepository(session).get(workspace_id)
-            assert workspace is not None
-            workspace.monitor_claimed_by = "worker-new"
-            await session.commit()
-
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        side_effect=_supersede_monitor_claim_after_restart,
-    )
-    compose_file = _write_compose_file(tmp_path)
-
-    with pytest.raises(_MonitorAgentServiceRecoverySupersededError):
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id=workspace_id,
-            compose_project="proj",
-            compose_file=compose_file,
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert adapter.calls == ["fix the comment"]
-    ensure_project_up.assert_awaited_once()
-
-
-@pytest.mark.unit
-async def test_monitor_agent_unrelated_cleanup_failure_is_not_recovered(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    cleanup_error = ComposeExecCleanupError(
-        invocation_id="awf-test-cleanup",
-        source="agent",
-        label="monitor",
-        message="permission denied",
-        cleanup_result=CommandResult(
-            returncode=1,
-            stdout="",
-            stderr="permission denied",
-        ),
-    )
-    adapter = FakeAdapter()
-    adapter.queue(exc=cleanup_error)
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=adapter,
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-    mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.probe_agent_service_health",
-        return_value=False,
-    )
-    ensure_project_up = mocker.patch(
-        "awf.runtime.pr_monitor_runner.agent_service_recovery.ComposeManager.ensure_project_up",
-        return_value=None,
-    )
-
-    with pytest.raises(ComposeExecCleanupError) as raised:
-        await runner._run_monitor_agent_with_service_recovery(
-            workspace_id="ws_monitor_cleanup_passthrough",
-            compose_project="proj",
-            compose_file=tmp_path / "compose.yml",
-            prompt="fix the comment",
-            log_source="recovery",
-            command_evidence=[],
-        )
-
-    assert raised.value is cleanup_error
-    ensure_project_up.assert_not_awaited()
