+++ tests/unit/service/test_worker.py
@@ -189,6 +189,7 @@ def test_build_worker_runtime_wires_executor_and_feature_monitor_factory(
             pr_monitor_factory: object,
             log_store: object,
             usage_sampler: object = None,
+            agent_runtime_executor: object = None,
         ) -> None:
             created["executor"] = self
             created["executor_session_factory"] = session_factory
